Reorganize services classes, OAuthKafkaPrincipal, and tests
