+++ oauth-common/src/main/java/io/strimzi/kafka/oauth/services/Services.java
@@ -2,7 +2,7 @@
  * Copyright 2017-2020, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.kafka.oauth.server.services;
+package io.strimzi.kafka.oauth.services;
 
 import java.util.Map;
 import java.util.concurrent.ExecutorService;
