+++ oauth-server/src/main/java/io/strimzi/kafka/oauth/server/JaasServerOauthValidatorCallbackHandler.java
@@ -8,8 +8,8 @@ import io.strimzi.kafka.oauth.common.Config;
 import io.strimzi.kafka.oauth.common.ConfigUtil;
 import io.strimzi.kafka.oauth.common.BearerTokenWithPayload;
 import io.strimzi.kafka.oauth.common.PrincipalExtractor;
-import io.strimzi.kafka.oauth.server.services.Services;
-import io.strimzi.kafka.oauth.server.services.ValidatorKey;
+import io.strimzi.kafka.oauth.services.Services;
+import io.strimzi.kafka.oauth.services.ValidatorKey;
 import io.strimzi.kafka.oauth.validator.JWTSignatureValidator;
 import io.strimzi.kafka.oauth.common.TokenInfo;
 import io.strimzi.kafka.oauth.validator.OAuthIntrospectionValidator;
