+++ oauth-server/src/main/java/io/strimzi/kafka/oauth/server/OAuthKafkaPrincipal.java
@@ -2,7 +2,7 @@
  * Copyright 2017-2019, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.kafka.oauth.server.services;
+package io.strimzi.kafka.oauth.server;
 
 import io.strimzi.kafka.oauth.common.BearerTokenWithPayload;
 import org.apache.kafka.common.security.auth.KafkaPrincipal;
@@ -12,7 +12,7 @@ import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import static io.strimzi.kafka.oauth.common.LogUtil.mask;
 
 /**
- * This class uses the KafkaPrincipal object to store additional info obtained at session authentication time,
+ * This class extends the KafkaPrincipal object to store additional info obtained at session authentication time,
  * and required later by a custom authorizer.
  *
  * Any additional fields should not be included in equals / hashcode check. If they are, that will break re-authentication.
@@ -37,7 +37,7 @@ public final class OAuthKafkaPrincipal extends KafkaPrincipal {
 
     @Override
     public String toString() {
-        return "JwtKafkaPrincipal(" + super.toString() + ", session:" + (jwt != null ? jwt.getSessionId() : "") +
+        return "OAuthKafkaPrincipal(" + super.toString() + ", session:" + (jwt != null ? jwt.getSessionId() : "") +
                 ", token:" + (jwt != null ? mask(jwt.value()) : "") + ")";
     }
 }
