+++ oauth-server/src/test/java/io/strimzi/kafka/oauth/server/MockBearerTokenWithPayload.java
@@ -2,7 +2,7 @@
  * Copyright 2017-2019, Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.kafka.oauth.server.authorizer;
+package io.strimzi.kafka.oauth.server;
 
 import io.strimzi.kafka.oauth.common.BearerTokenWithPayload;
